// src/api/back/models/DropSettingsDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DropSettingsDto
 */
export interface DropSettingsDto {
    /**
     * 
     * @type {number}
     * @memberof DropSettingsDto
     */
    baseDropChance: number;
    /**
     * 
     * @type {number}
     * @memberof DropSettingsDto
     */
    subsequentDropChance: number;
    /**
     * 
     * @type {number}
     * @memberof DropSettingsDto
     */
    desiredStock: number;
}

export function DropSettingsDtoFromJSON(json: any): DropSettingsDto {
    return DropSettingsDtoFromJSONTyped(json, false);
}

export function DropSettingsDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): DropSettingsDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'baseDropChance': json['baseDropChance'],
        'subsequentDropChance': json['subsequentDropChance'],
        'desiredStock': json['desiredStock'],
    };
}

export function DropSettingsDtoToJSON(value?: DropSettingsDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'baseDropChance': value.baseDropChance,
        'subsequentDropChance': value.subsequentDropChance,
        'desiredStock': value.desiredStock,
    };
}