// src/api/back/models/DropTierDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DropTierDto
 */
export interface DropTierDto {
    /**
     * 
     * @type {number}
     * @memberof DropTierDto
     */
    minPrice: number;
    /**
     * 
     * @type {number}
     * @memberof DropTierDto
     */
    maxPrice: number;
    /**
     * 
     * @type {number}
     * @memberof DropTierDto
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof DropTierDto
     */
    weight: number;
}

export function DropTierDtoFromJSON(json: any): DropTierDto {
    return DropTierDtoFromJSONTyped(json, false);
}

export function DropTierDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): DropTierDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'minPrice': json['minPrice'],
        'maxPrice': json['maxPrice'],
        'id': json['id'],
        'weight': json['weight'],
    };
}

export function DropTierDtoToJSON(value?: DropTierDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'minPrice': value.minPrice,
        'maxPrice': value.maxPrice,
        'id': value.id,
        'weight': value.weight,
    };
}