// src/api/back/models/UpdateDropTierDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateDropTierDto
 */
export interface UpdateDropTierDto {
    /**
     * 
     * @type {number}
     * @memberof UpdateDropTierDto
     */
    minPrice?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateDropTierDto
     */
    maxPrice?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateDropTierDto
     */
    weight?: number;
}

export function UpdateDropTierDtoFromJSON(json: any): UpdateDropTierDto {
    return UpdateDropTierDtoFromJSONTyped(json, false);
}

export function UpdateDropTierDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateDropTierDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'minPrice': !exists(json, 'minPrice') ? undefined : json['minPrice'],
        'maxPrice': !exists(json, 'maxPrice') ? undefined : json['maxPrice'],
        'weight': !exists(json, 'weight') ? undefined : json['weight'],
    };
}

export function UpdateDropTierDtoToJSON(value?: UpdateDropTierDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'minPrice': value.minPrice,
        'maxPrice': value.maxPrice,
        'weight': value.weight,
    };
}